requirements.txt
click
duckdb
fastapi
polars
pyarrow
pyyaml
uvicorn
